+++ flowman-kernel-server/src/main/scala/com/dimajix/flowman/tools/kernel/Kernel.scala
@@ -41,7 +41,6 @@ import com.dimajix.flowman.kernel.service.MultiWorkspaceManager
 import com.dimajix.flowman.kernel.service.WorkspaceManager
 import com.dimajix.hadoop.HADOOP_VERSION
 import com.dimajix.spark.SPARK_VERSION
-import com.dimajix.spark.features.isDatabricks
 
 
 
@@ -149,9 +148,7 @@ final class Kernel(args:Arguments) extends Tool {
         if (!server.isTerminated()) {
             server.stop()
 
-            // github-357: Spark session should not be shut down in DataBricks environment
-            if (!isDatabricks)
-                rootSession.shutdown()
+            rootSession.shutdown()
         }
     }