+++ flowman-tools/src/main/scala/com/dimajix/flowman/tools/exec/Driver.scala
@@ -37,7 +37,6 @@ import com.dimajix.flowman.execution.Status
 import com.dimajix.flowman.model.Project
 import com.dimajix.hadoop.HADOOP_VERSION
 import com.dimajix.spark.SPARK_VERSION
-import com.dimajix.spark.features.isDatabricks
 
 
 object Driver {
@@ -149,9 +148,7 @@ class Driver(options:Arguments) extends Tool {
 
         val result = options.command.execute(session, project, context)
 
-        // github-357: Spark session should not be shut down in DataBricks environment
-        if (!isDatabricks)
-            session.shutdown()
+        session.shutdown()
 
         result
     }